requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
openai==1.7.1
colorama==0.4.4